Tweak parameters
Accuracy has now jumped up.

_Bezos, for example, is able to be detected from his side profile with 40% accuracy_

Masks seem to _sorta_ work. _Musk is seen with 36% accuracy, but the one masked photo of Bezos is classified as Devito_

Detected faces (but not classified) are now labeled with the "closest face"
The punch time stamp is grab upon program startup, rather than after loading all libraries and running the face through the model.