Hacky inclusion of multiple employee faces.
With the previous version, when adding the face encoding of a particular employee to the `known_encodings` dictionary, we would use the employee's ID as the key.  In turn, this would mean only one face encoding was saved per employee (as the employee would only have one ID).  A hacky way of getting around this was to append a "run" integer to the ID (as so: `<ID> + ":" + <run>`) to make each sure each encoding is in the dictionary but still retains the employee's ID.  The problem with this implementation is that now each encoding of an employee is save to a unique key and essentially make each of their encodings a new "cluster".  So rather than cluster by employee ID, we're clustering by image.
To work with this, when we find the closest "match" for classification we split the key (remember, this key is `<ID>:<run>`) and grab the front of that split (we grab the ID from <ID>:<run>)

In the end matches have been boosted by ~3%. Only really tested this claim on one photo, but it be what it be.

Also added the `-ns` or `-noshow` parameter to `-punch`, which will NOT display the image showing the employee being identified.  This image is still generated, just not shown. (Don't really need it anymore, as I know faces can be found correctly now...)