2 Layer Encoding Model
Vision: Fix way encodings are saved to disk (as csv's)

face_recognition: Rearrange the way the models are loaded to improve import speed

encodingNet: The machine learning model for encodings. We take the 128d encoding and feed it into the model produced by this file to classify the 128d input encoding as an employee's ID. A little overkill since we _could_ use K-Nearest, but this adds some ML model to the project. Additionally, it opens the door for improved accuracy and clustering (of employee's encodings).
encodingNet loads the encodings database generated by the "training" function in vision.py. These encodings are then put into feature_set and output labels are generated based on the employee's ID for each encoding.

We then create a 2 layer neural network (or 3 layer, if so desired) with randomized biases and weights, then we run feedforward() and backprop() to update the weights when needed.
Currently this is done `epochs` times (500000) to generate a "trained" model. We can then save the weights and biases of the model to a file for portability. Initial testing yields a % accuracy of around 12.5 on a dataset of 8 (8 encodings across 3 employees).
Accuracy is low, but that's likely due to the size of the training dataset.