Final code upload
This was the last modifications made to VisionNet before the Vision paper was submitted.

generateVisionNet.py: A bunch of model testing and logging calls (generating data for paper)

visionAPI: Draw dots on each face landmark (for visualization in paper)
The original box drawing code is still there (commented out).

No more updates will be posted to this branch.